Switch from `rand` to `fastrand`
This crate's need for randomness is for the jitter feature only, and so
does not need to be cryptographically secure. As such, the lighter-weight
`fastrand` crate can be used instead of `rand` to reduce the number of
transitive dependencies and impact on compile times. (See the diff
of `Cargo.lock` for how noticeable a difference this makes.)

The `rand` crate itself mentions `fastrand` as a simpler alternative:
https://github.com/rust-random/rand/blob/master/README.md

And `fastrand` is also used by several other popular crates - so for
some projects, will already be in their dependency tree:
https://crates.io/crates/fastrand/reverse_dependencies

See:
https://crates.io/crates/fastrand
https://docs.rs/fastrand/latest/fastrand/